Generate validation schema for CLN
