+++ lib/validate.js
@@ -6,11 +6,10 @@ import {
 } from './constants'
 import { SUPPORTED_CURRENCIES } from './currency'
 import { NOSTR_MAX_RELAY_NUM, NOSTR_PUBKEY_BECH32, NOSTR_PUBKEY_HEX } from './nostr'
-import { msatsToSats, numWithUnits, abbrNum, ensureB64, B64_URL_REGEX } from './format'
+import { msatsToSats, numWithUnits, abbrNum, ensureB64 } from './format'
 import * as usersFragments from '@/fragments/users'
 import * as subsFragments from '@/fragments/subs'
 import { datePivot } from './time'
-import { decodeRune } from '@/lib/cln'
 import bip39Words from './bip39-words'
 
 const { SUB } = subsFragments
@@ -300,32 +299,6 @@ export function subSelectSchemaMembers (args) {
 export function advSchema (args) {
   return object({
     ...advPostSchemaMembers(args)
-  })
-}
-
-export function CLNAutowithdrawSchema ({ me } = {}) {
-  return object({
-    socket: string().socket().required('required'),
-    rune: string().matches(B64_URL_REGEX, { message: 'invalid rune' }).required('required')
-      .test({
-        name: 'rune',
-        test: (v, context) => {
-          const decoded = decodeRune(v)
-          if (!decoded) return context.createError({ message: 'invalid rune' })
-          if (decoded.restrictions.length === 0) {
-            return context.createError({ message: 'rune must be restricted to method=invoice' })
-          }
-          if (decoded.restrictions.length !== 1 || decoded.restrictions[0].alternatives.length !== 1) {
-            return context.createError({ message: 'rune must be restricted to method=invoice only' })
-          }
-          if (decoded.restrictions[0].alternatives[0] !== 'method=invoice') {
-            return context.createError({ message: 'rune must be restricted to method=invoice only' })
-          }
-          return true
-        }
-      }),
-    cert: hexOrBase64Validator,
-    ...autowithdrawSchemaMembers
   })
 }
