+++ lib/wallet.js
@@ -1,6 +1,7 @@
 import { array, object, string } from 'yup'
 import { autowithdrawSchemaMembers, hexOrBase64Validator, lightningAddressValidator } from '@/lib/validate'
 import { TOR_REGEXP } from '@/lib/url'
+import { B64_URL_REGEX } from '@/lib/format'
 
 export function generateResolverName (walletField) {
   const capitalized = walletField[0].toUpperCase() + walletField.slice(1)
@@ -76,6 +77,8 @@ export function generateSchema (wallet) {
     if (validationType === 'socket') validator = string().socket()
 
     if (validationType === 'hexOrBase64') validator = hexOrBase64Validator
+
+    if (validationType === 'b64url') validator = string().matches(B64_URL_REGEX, { message: `invalid ${field.name}` })
 
     if (min !== undefined) validator = validator.min(min)
     if (max !== undefined) validator = validator.max(max)
